fix: changes after code review
