+++ packages/edition-slices/src/tiles/tile-s/styles/index.js
@@ -1,10 +1,10 @@
 import {
   fonts,
+  editionBreakpoints,
   spacing,
   fontSizes,
   colours
 } from "@times-components/styleguide";
-import { verticalStyles } from "../../shared/styles";
 
 const styles = {
   bold: {
@@ -26,7 +26,6 @@ const styles = {
     fontSize: 14,
     lineHeight: 20
   },
-  star: verticalStyles,
   title: {
     color: colours.functional.brandColour,
     fontFamily: fonts.headline,
@@ -38,4 +37,38 @@ const styles = {
   }
 };
 
-export default styles;
+const mediumBreakpointStyles = {
+  bold: {
+    color: colours.functional.brandColour,
+    fontWeight: "bold"
+  },
+  byline: {
+    color: colours.functional.secondary,
+    fontFamily: fonts.supporting,
+    fontSize: 12,
+    marginTop: spacing(2)
+  },
+  container: {
+    padding: spacing(3)
+  },
+  paragraph: {
+    color: colours.functional.secondary,
+    fontFamily: fonts.bodyRegular,
+    fontSize: 14,
+    lineHeight: 20
+  },
+  title: {
+    color: colours.functional.brandColour,
+    fontFamily: fonts.headline,
+    fontSize: 20,
+    lineHeight: 20
+  },
+  titleWrapper: {
+    flexDirection: "row",
+    alignItems: "center",
+    marginBottom: spacing(2)
+  }
+};
+
+export default breakpoint =>
+  breakpoint === editionBreakpoints.medium ? mediumBreakpointStyles : styles;
